refactor(core): put modules into a standalone dir
