+++ packages/core/src/index.ts
@@ -1,5 +1,4 @@
-export * from "./auth";
-export * from "./catalog";
+export * from "./modules";
 export * from "./search";
 export * from "./utils";
 export * from "./types";
