+++ packages/core/src/internal.ts
@@ -1,6 +1,5 @@
-export * from "./catalog";
+export * from "./modules";
 export * from "./search";
-export * from "./auth";
 export * from "./types";
 export * from "./utils";
 export * from "./error";
