+++ packages/core/src/modules/auth/index.ts
