+++ packages/core/src/modules/catalog/song/container.ts
@@ -1,7 +1,7 @@
 import { prisma } from "@cvsa/db";
 import { SongRepository } from "./repository";
 import { SongService } from "./service";
-import { SongSearchService, searchManager } from "../../search";
+import { SongSearchService, searchManager } from "../../../search";
 import type { EmbeddingApp } from "@cvsa/embedding";
 import { treaty } from "@elysiajs/eden";
 export const embeddingManager = treaty<EmbeddingApp>("localhost:14900");
