+++ packages/core/src/modules/catalog/song/index.ts
