+++ packages/core/src/modules/index.ts
@@ -1,0 +1,2 @@
+export * from "./catalog";
+export * from "./auth";
